requirements.txt
Flask
SQLAlchemy
psycopg2-binary
python-dotenv
flask_sqlalchemy
flask-login
steam-openid
requests